zigbee-infra/infra/control-api/pyproject.toml
[project]
name = "zigbee-control-api"
version = "0.1.0"
description = "Minimal control API for Zigbee demo"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.27.1",
  "paho-mqtt>=2.1.0",
]

[project.scripts]
control-api = "control_api.main:run"
